% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrev_state.R
\name{abrev_state}
\alias{abrev_state}
\title{AbbreviateUS State Names}
\usage{
abrev_state(state, rm_nums = FALSE)
}
\arguments{
\item{state}{A vector of full state names}
}
\value{
A vector of 2 letter state abbreviations
}
\description{
AbbreviateUS State Names
}
\examples{
abrev_state(c("Vermont", "District of Columbia", "France", "Maine 2nd"), rm_nums = TRUE)
}
