% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campfin-package.R
\docType{package}
\name{campfin-package}
\alias{campfin}
\alias{campfin-package}
\title{\code{campfin} package}
\description{
The campfin package was written to facilitate the work done on The
Accountability Project (TAP) by the data journalists at The Investigative
Reporting Workshop in Washington, DC.
}
\details{
TAP is an effort to centralize many public databases into a searchable
website. To improve the searchability of the database, the campfin package
is used to normalize values in a common format. The normalization vignette
provides examples on how this process improved a database.

The other functions in this package are written to facilitate the exploration
of a database. The Investigative Reporting Workshop writes public data
diaries to document the data wrangling process; the open source campfin
functions help download, import, explore, and wrangle public database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/irworkshop/campfin}
  \item \url{https://irworkshop.github.io/campfin/}
  \item Report bugs at \url{https://github.com/irworkshop/campfin/issues}
}

}
\author{
\strong{Maintainer}: Kiernan Nicholls \email{kiernann@protonmail.com} [copyright holder]

Authors:
\itemize{
  \item Yanqi Xu \email{yanqixu@outlook.com}
}

Other contributors:
\itemize{
  \item Investigative Reporting Workshop [copyright holder]
  \item Schuyler Erle \email{schuyler@geocoder.us} [copyright holder]
}

}
\keyword{internal}
