% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campfin.R
\docType{package}
\name{campfin}
\alias{campfin}
\alias{campfin-package}
\title{\code{campfin} package}
\description{
Wrangling Campaign Finance
}
\details{
See the README on \href{https://github.com/kiernann/campfin/#readme}{GitHub}
}
