% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_date_usa.R
\name{col_date_usa}
\alias{col_date_usa}
\title{Parse USA (MDY) Dates}
\usage{
col_date_usa()
}
\value{
A \code{POSIXct()} vector
}
\description{
Parse MM/DD/YYY dates
}
\examples{
readr::read_csv(file = "x\\n09/18/1996", col_types = readr::cols(x = col_date_usa()))
}
