% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{count_diff}
\alias{count_diff}
\title{Count Set Difference}
\usage{
count_diff(x, y)
}
\arguments{
\item{x}{A vector to check.}

\item{y}{A vector to compare against.}
}
\value{
The number of \emph{unique} values of \code{x} not in \code{y}.
}
\description{
Find length of the set of differences between \code{x} and \code{y}.
}
\details{
\code{sum(x \%out\% y)}
}
\examples{
# Only unique values are checked
count_diff(c("VT", "NH", "ZZ", "ZZ", "ME"), state.abb)
}
\seealso{
Other Simple Counting Wrappers: \code{\link{count_in}},
  \code{\link{count_na}}, \code{\link{count_out}},
  \code{\link{progress_table}}, \code{\link{prop_in}},
  \code{\link{prop_na}}, \code{\link{prop_out}}
}
\concept{Simple Counting Wrappers}
