% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_fun.R
\name{count_na}
\alias{count_na}
\title{Count NA Values}
\usage{
count_na(x)
}
\arguments{
\item{x}{A vector to count \code{NA}}
}
\value{
The number of \code{NA} values in a vector
}
\description{
Wrap around \code{sum(is.na())}
}
\examples{
count_na(dplyr::starwars$gender)
}
