% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{count_out}
\alias{count_out}
\title{Count out}
\usage{
count_out(x, y, na.rm = TRUE)
}
\arguments{
\item{x}{A vector to check.}

\item{y}{A vector to compare against.}

\item{na.rm}{logical; Should \code{NA} be ignored?}
}
\value{
The sum of \code{x} absent in \code{y}.
}
\description{
Count the total values of \code{x} that are are \verb{\%out\%} of the vector \code{y}.
}
\details{
\code{sum(x \%out\% y)}
}
\examples{
count_out(c("VT", "NH", "ZZ", "ME"), state.abb)
}
\seealso{
Other Simple counting wrappers: 
\code{\link{count_diff}()},
\code{\link{count_in}()},
\code{\link{count_na}()},
\code{\link{prop_in}()},
\code{\link{prop_na}()},
\code{\link{prop_out}()}
}
\concept{Simple counting wrappers}
