% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{count_vec}
\alias{count_vec}
\title{Count Values of a Vector}
\usage{
count_vec(x, sort = TRUE)
}
\arguments{
\item{x}{A vector to check.}

\item{sort}{logical; if TRUE will sort output in descending order of \code{n}}
}
\value{
A tibble, with counts of each \code{value} in \code{n}.
}
\description{
A version of \code{\link[dplyr:count]{dplyr::count()}} which uses \code{\link[tibble:enframe]{tibble::enframe()}} to
count the number of values in a single vector.
}
\concept{Simple Counting Wrappers
count_vec(x = rivers)
count_vec(x = sample(x = state.name, size = 1000, replace = TRUE))}
