% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dark2}
\alias{dark2}
\title{Dark Color Palette}
\format{
A named character vector of hex color codes (length 8).
}
\usage{
dark2
}
\description{
The Dark2 brewer color palette
}
\keyword{datasets}
