% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-time.R
\name{download_time}
\alias{download_time}
\title{Calculate the time to download a file}
\usage{
download_time(x, speed = NULL)
}
\arguments{
\item{speed}{The bandwidth in megabytes per second (Mbps) of the internet
connection. If not provided, the function will call
\code{\link[speedtest:spd_download_test]{speedtest::spd_download_test()}} on the best availber servers and return
the median bandwidth of the test.}
}
\value{
The number of seconds needed to download a given file.
}
\description{
Using \code{\link[httr:HEAD]{httr::HEAD()}} and \code{\link[speedtest:spd_test]{speedtest::spd_test()}} to check the
size of a file before downloading.
}
\examples{
stats median
}
