% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_abbrev.R
\name{expand_abbrev}
\alias{expand_abbrev}
\title{Expand Abbreviations}
\usage{
expand_abbrev(x, abb = NULL, rep = NULL)
}
\arguments{
\item{city}{A vector of city names.}

\item{geo_abbs}{A two-column data frame like \link{usps_city}, with a full
geographic feature ("LAKE") in the first column and abbreviations ("LK")
in the second. Replace all abbreviations with their full version.}

\item{st_abbs}{A vector of state abbreviations ("VT") to remove from the
\emph{end} (and only end) of city names ("STOWE VT").}

\item{na}{A vector of values to make \code{NA} (useful with the \link{invalid_city}
vector).}

\item{na_rep}{logical; If \code{TRUE}, replace all single digit (repeating)
strings with \code{NA}.}
}
\value{
A vector of normalized city names.
}
\description{
Create or use a named vector (\code{c(abb = rep)}) and pass it to
\code{\link[stringr:str_replace_all]{stringr::str_replace_all()}}
}
\examples{
expand_abbrev(x = "MT VERNON", abb = c("MT" = "MOUNT"))
}
