% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extra_city}
\alias{extra_city}
\title{Additional US City Names}
\format{A sorted vector of unique locality names (length 127).}
\usage{
extra_city
}
\description{
Cities not contained in \link{valid_city}, but are
accepted localities (neighborhoods or census designated
places).
}
\keyword{datasets}
