% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-age.R
\name{file_age}
\alias{file_age}
\title{File modification date age}
\usage{
file_age(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=file.info]{file.info()}}, namely character vectors
containing file paths. Tilde-expansion is done: see \code{\link[=path.expand]{path.expand()}}.}
}
\value{
A Period class object.
}
\description{
The period of time since a system file was modified.
}
\examples{
file_age(system.file("README.md", package = "campfin"))
}
