% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-encoding.R
\name{file_encoding}
\alias{file_encoding}
\title{File Encoding}
\usage{
file_encoding(path)
}
\arguments{
\item{path}{A local file path or glob to check.}
}
\value{
A tibble of file encoding.
}
\description{
Call the \code{file} command line tool with option \code{-i}.
}
