% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_dupes.R
\name{flag_dupes}
\alias{flag_dupes}
\title{Flag Duplicate Rows}
\usage{
flag_dupes(df, ...)
}
\arguments{
\item{df}{A data frame}

\item{...}{Arguments to pass to dplyr::select() (e.g., dplyr::everything())}
}
\value{
A data frame with a new \code{dupe_flag} logical variable
}
\description{
Flag Duplicate Rows
}
\examples{
flag_dupes(iris, dplyr::everything())
}
