% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_na.R
\name{flag_na}
\alias{flag_na}
\title{Flag Missing Values}
\usage{
flag_na(data, ...)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame with a new \code{na_flag} logical variable
}
\description{
Flag Missing Values
}
\examples{
flag_na(dplyr::storms, ts_diameter)
}
