% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flush-memory.R
\name{flush_memory}
\alias{flush_memory}
\title{Flush Garbage Memory}
\usage{
flush_memory(n = 1)
}
\arguments{
\item{n}{The number of times to run \code{\link[=gc]{gc()}}.}
}
\description{
Run a full \code{\link[=gc]{gc()}} a number of times.
}
