% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_fun.R
\name{glimpse_fun}
\alias{glimpse_fun}
\title{Check for All New Files}
\usage{
glimpse_fun(data, fun)
}
\arguments{
\item{data}{A data frame to glimpse.}

\item{fun}{A function to map to each column.}
}
\value{
A tibble with a row for every column and new columns with count and proportion.
}
\description{
Check for All New Files
}
\examples{
glimpse_fun(dplyr::storms, dplyr::n_distinct)
}
