% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{\%out\%}
\alias{\%out\%}
\title{Inverted Match}
\usage{
x %out% y
}
\arguments{
\item{x}{A vector to check.}

\item{y}{A vector to compare against.}
}
\value{
Logical if X is not present in Y
}
\description{
Inverted Match
}
\examples{
c("A", "B", "0") \%out\% LETTERS
}
