% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess-delims.R
\name{guess_delim}
\alias{guess_delim}
\title{Guess the delimiter of a text file}
\source{
\url{https://github.com/r-lib/vroom/blob/7b7e775fd6d4261532116bcc881879f9439f29df/R/vroom.R}
}
\usage{
guess_delim(path, delims = c(",", "\\t", " ", "|", ":", ";"))
}
\arguments{
\item{path}{The path of a delimited file to read.}

\item{delims}{The single characters to guess from.}
}
\value{
The single character guessed as a delimiter.
}
\description{
Taken from code used in \code{\link[vroom:vroom]{vroom::vroom()}} with automatic reading.
}
\examples{
guess_delim(system.file("extdata", "vt_contribs.csv", package = "campfin"))
}
