% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-filename.R
\name{http_filename}
\alias{http_filename}
\title{Extract a content file name from HTTP request headers}
\usage{
http_filename(x)
}
\arguments{
\item{x}{An \code{\link[httr:GET]{httr::GET()}} request object.}
}
\value{
File name character string
}
\description{
Extract a content file name from HTTP request headers
}
