% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invalid_city}
\alias{invalid_city}
\title{Invalid City Names}
\format{
A vector of length 54.
}
\usage{
invalid_city
}
\description{
A custom vector containing common invalid city names.
}
\keyword{datasets}
