% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_abbrev.R
\name{is_abbrev}
\alias{is_abbrev}
\title{Check if Abbreviation}
\usage{
is_abbrev(abb, full)
}
\arguments{
\item{abb}{A suspected abbreviation}

\item{full}{A long form string to test against}
}
\value{
A logical if \code{abb} is potential abbreviation of \code{full}
}
\description{
Check if Abbreviation
}
\examples{
is_abbrev(abb = "BRX", full = "BRONX")
is_abbrev(abb = "BRX", full = "BROOKLYN")
}
