% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-even.R
\name{is_even}
\alias{is_even}
\title{Check if Even}
\usage{
is_even(x)
}
\arguments{
\item{x}{An integer.}
}
\value{
logical; Whether the integer is even or odd.
}
\description{
Is a number even? Pretty simple...
}
\examples{
is_even(5)
is_even(10L)
}
