% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keypad_convert.R
\name{keypad_letters}
\alias{keypad_letters}
\title{Return the Keypad Letters for a Number}
\usage{
keypad_letters(number)
}
\arguments{
\item{number}{A \emph{single} number.}
}
\value{
The numbers which would return \code{number} in \code{\link[=keypad_number]{keypad_number()}}.
}
\description{
Return the Keypad Letters for a Number
}
