% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keypad_convert.R
\name{keypad_number}
\alias{keypad_number}
\title{Return the Keypad Number for a Letter}
\usage{
keypad_number(letter)
}
\arguments{
\item{letter}{A character vector of letters.}
}
\value{
A character vector of numbers.
}
\description{
Return the Keypad Number for a Letter
}
\examples{
keypad_number("Example")
}
