% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_common.R
\name{most_common}
\alias{most_common}
\title{Most Common Values}
\usage{
most_common(x, n = 6)
}
\arguments{
\item{x}{A vector}

\item{n}{number of values to return}
}
\value{
Sorted vector of \code{n} most common values
}
\description{
Most Common Values
}
\examples{
most_common(dplyr::storms$month, 3)
}
