% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{na_rep}
\alias{na_rep}
\title{Remove repeated character elements}
\usage{
na_rep(x, n = 1)
}
\arguments{
\item{x}{A vector to check.}

\item{n}{The minumum number times a character must repeat.}
}
\value{
The vector \code{x} with \code{NA} replacing repeating character values.
}
\description{
Set \code{NA} for the values of \code{x} that contain a single repeating character.
}
\details{
Uses the regular expression \code{"^(.)\\1+$"}.
}
\concept{Simple counting wrappers
na_rep(c("VT", "NH", "ZZ", "ME"))}
