% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-ascii.R
\name{non_ascii}
\alias{non_ascii}
\title{Show non-ASCII lines of file}
\usage{
non_ascii(path, highlight = FALSE)
}
\arguments{
\item{path}{The path to a text file to check.}

\item{highlight}{A function used to add ANSI escapes to highlight bytes.}
}
\value{
Tibble of line locations.
}
\description{
Show non-ASCII lines of file
}
\examples{
non_ascii(system.file("README.md", package = "campfin"))
}
