% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_address.R
\name{normal_address}
\alias{normal_address}
\title{Normalize US Street Addresses}
\usage{
normal_address(address, add_abbs = NULL, na = c("", "NA"),
  na_rep = FALSE)
}
\arguments{
\item{address}{A vector of street addresses (ideally without city, state, and ZIP).}

\item{add_abbs}{A two-column data frame with address abbreviations ("ST") in the first column and replacement strings ("STREET") in the second.}

\item{na}{A vector of values to make \code{NA}.}

\item{na_rep}{If \code{TRUE}, make all single digit repeating strings \code{NA}.}
}
\value{
A vector of normalized street addresses.
}
\description{
Normalize US Street Addresses
}
