% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_phone.R
\name{normal_phone}
\alias{normal_phone}
\title{Normalize a US Phone Number}
\usage{
normal_phone(number, format = "(\%a) \%e-\%l", na_bad = FALSE,
  na_rep = FALSE, convert = FALSE, rm_ext = FALSE)
}
\arguments{
\item{number}{A vector of phone number in any format.}

\item{format}{The desired output format, with \code{\%a} representing the 3-digit
area code, \code{\%e} representing the 3-digit exchange, and \code{\%l} representing
the 4-digit line number.}

\item{na_bad}{logical; Should invalid numbers be replaced with \code{NA}.}

\item{na_rep}{logical; Should single digit (repeating) strings be replaced
with \code{NA}.}

\item{convert}{logical; Should LETTERS be converted to their keypad number
equivalent.}

\item{rm_ext}{logical; Should extensions be removed from the end of a number.}
}
\value{
A normalized telephone number.
}
\description{
Take US phone numbers in any number of formats and try to
convert them to a standard format.
}
\examples{
normal_phone(number = c("916-225-5887"))
}
