% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_zip.R
\name{normal_zip}
\alias{normal_zip}
\title{Normalize US ZIP Codes}
\usage{
normal_zip(zip, na = c(""), na_rep = FALSE)
}
\arguments{
\item{zip}{A vector ostensibly of US ZIP codes}

\item{na}{A vector of values to make `NA`}

\item{na_rep}{If TRUE, make all single digit repeating strings `NA`}
}
\value{
A vector of normalized 5-digit ZIP codes
}
\description{
Normalize US ZIP Codes
}
