% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_table.R
\name{progress_table}
\alias{progress_table}
\title{Create a Progress Table}
\usage{
progress_table(..., compare)
}
\arguments{
\item{...}{Any number of vectors to check.}

\item{compare}{A vector to compare each of \code{...} against. Useful
with \link{valid_zip}, \link{valid_state} (\link{valid_name}), or
\link{valid_city}.}
}
\value{
A table with a row for each vector in \code{...}.
}
\description{
Create a tibble with rows for each stage of normalization
and columsn for the various statistics most useful in assessing
the progress of each stage.
}
\seealso{
Other Simple Counting Wrappers: \code{\link{count_in}},
  \code{\link{count_na}}, \code{\link{count_out}},
  \code{\link{prop_in}}, \code{\link{prop_na}},
  \code{\link{prop_out}}
}
\concept{Simple Counting Wrappers}
