% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{prop_na}
\alias{prop_na}
\title{Proportion missing}
\usage{
prop_na(x)
}
\arguments{
\item{x}{A vector to check.}
}
\value{
The sum of \code{x} absent in \code{y}.
}
\description{
Find the proportion of values of \code{x} that are \code{NA}.
}
\details{
\code{mean(is.na(x))}
}
\examples{
prop_na(c("VT", "NH", NA, "ME"))
}
\seealso{
Other Simple counting wrappers: 
\code{\link{count_diff}()},
\code{\link{count_in}()},
\code{\link{count_na}()},
\code{\link{count_out}()},
\code{\link{prop_in}()},
\code{\link{prop_out}()}
}
\concept{Simple counting wrappers}
