% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-names.R
\name{read_names}
\alias{read_names}
\title{Read column names}
\usage{
read_names(file, delim = guess_delim(file))
}
\arguments{
\item{file}{Path to text file.}

\item{delim}{Character separating column names.}
}
\value{
Character vector of column names.
}
\description{
Read the first line of a delimited file as vector.
}
\examples{
read_names("date,lgl\n11/09/2016,TRUE")
}
