% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx-break.R
\name{rx_break}
\alias{rx_break}
\title{Form a Word Break Regex}
\usage{
rx_break(pattern)
}
\arguments{
\item{pattern}{A regex pattern (a word) to wrap in \code{\\b}.}
}
\value{
The \code{pattern} wrapped in \code{\\b}.
}
\description{
Wrape a word in word boundry (\code{\\b}) characters. Useful when
combined with \code{\link[stringr:str_which]{stringr::str_which()}} and \code{\link[stringr:str_detect]{stringr::str_detect()}} to match
only entire words and not that word \emph{inside} another word (e.g., "sting"
and "testing").
}
\examples{
rx_break("test")
rx_break(state.abb)
}
