% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rx_phone}
\alias{rx_phone}
\title{Phone number regex}
\format{
A character string (length 1).
}
\usage{
rx_phone
}
\description{
The regex string to match US phone numbers in a variety of
common formats.
}
\keyword{datasets}
