% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rx_zip}
\alias{rx_zip}
\title{ZIP code regex}
\format{A single character string.}
\usage{
rx_zip
}
\description{
The regex string to extract 5 or 9 digit zip from end of address.
}
\keyword{datasets}
