% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_dist.R
\name{str_dist}
\alias{str_dist}
\title{String Distance}
\usage{
str_dist(a, b, method = "osa", ...)
}
\arguments{
\item{a}{\code{R} object (target); will be converted by \code{\link[base:as.character]{base::as.character()}}}

\item{method}{Method for distance calculation. The default is "osa."}

\item{...}{Other rguments passed to \code{\link[stringdist:stringdist]{stringdist::stringdist()}}.}

\item{a}{\code{R} object (source); will be converted by \code{\link[base:as.character]{base::as.character()}}}
}
\value{
The distance between string \code{a} and string \code{b}.
}
\description{
\code{str_dist} calls wraps around the
}
\examples{
str_dist(a = "BRULINGTON", b = "BURLINGTON")
}
