% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-normal.R
\name{str_normal}
\alias{str_normal}
\title{Normalize a character string}
\usage{
str_normal(x)
}
\arguments{
\item{x}{A character string to normalize.}
}
\value{
A normalized vector of the same length.
}
\description{
The generic normalization that underpins functions like \code{\link[=normal_city]{normal_city()}} and
\code{\link[=normal_address]{normal_address()}}. This function simply chains together three
\code{stringr::str_*()} functions:
\enumerate{
\item Convert to uppercase.
\item Replace punctutations with whitespaces.
\item Trim and squish excess whitespace.
}
}
\examples{
str_normal("   TestING 123   example_test.String   ")
}
\seealso{
Other geographic normalization functions: \code{\link{abbrev_full}},
  \code{\link{abbrev_state}}, \code{\link{check_city}},
  \code{\link{expand_abbrev}}, \code{\link{expand_state}},
  \code{\link{fetch_city}}, \code{\link{normal_address}},
  \code{\link{normal_city}}, \code{\link{normal_state}},
  \code{\link{normal_zip}}
}
\concept{geographic normalization functions}
