% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_files_new.R
\name{this_file_new}
\alias{this_file_new}
\title{Check if a File is New}
\usage{
this_file_new(path)
}
\arguments{
\item{path}{The path to a file to check.}
}
\value{
A logical value indicating whether or a file has modification date equal to today.
}
\description{
Check if a File is New
}
