% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url2path.R
\name{url2path}
\alias{url2path}
\title{Make a File Path from a URL}
\usage{
url2path(url, dir)
}
\arguments{
\item{url}{The URL of a file to download.}

\item{dir}{The directory where the file will be downloaded.}
}
\value{
The desired file path to a URL file.
}
\description{
Combine the \code{\link[=basename]{basename()}} of a file URL with a directory path.
}
\details{
Useful in the \code{destfile} argument to \code{\link[=download.file]{download.file()}} to save a file with
the same name as the URL's file name.
}
\examples{
url2path("https://floridalobbyist.gov/reports/llob.txt", tempdir())
}
