% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_file_size.R
\name{url_file_size}
\alias{url_file_size}
\title{Check URL File Size}
\usage{
url_file_size(url, format = FALSE)
}
\arguments{
\item{url}{The url of the page to retrieve}

\item{format}{Whether to format as byte measurment and symbol (e.g., Mb)}
}
\value{
The size of a file to be downloaded
}
\description{
Check URL File Size
}
\examples{
url_file_size("https://campaignfinance.cdn.sos.ca.gov/dbwebexport.zip", format = TRUE)
}
