% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usps}
\alias{usps}
\title{US City, State, and ZIPs}
\format{A tibble with 44,360 rows of 3 variables:
\describe{
\item{abb}{Commonly Used Street Suffix or Abbreviation}
\item{full}{Primary Street Suffix}
...
}}
\source{
The USPS C1 Street Suffix Abbreviations
\url{https://pe.usps.com/text/pub28/28apc_002.htm}
}
\usage{
usps
}
\description{
A dataset containing common street suffixes or suffix abbreviations and their full equivalent.
}
\keyword{datasets}
