% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usps_city}
\alias{usps_city}
\title{USPS C1 City Abbreviations}
\format{A tibble with 44,360 rows of 3 variables:
\describe{
\item{abb}{Commonly Used Street Suffix or Abbreviation}
\item{full}{Primary Street Suffix}
...
}}
\source{
The USPS C1 Street Suffix Abbreviations
\url{https://pe.usps.com/text/pub28/28apc_002.htm}
}
\usage{
usps_city
}
\description{
A currated and edited dataset containing the USPS abbreviations for only city names.
}
\keyword{datasets}
