% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usps_street}
\alias{usps_street}
\title{USPS Street Abbreviations}
\format{
A tibble with 325 rows of 3 variables:
\describe{
\item{full}{Primary Street Suffix.}
\item{abb}{Commonly Used Street Suffix or Abbreviation.}
...
}
}
\source{
USPS Appendix
\href{https://pe.usps.com/text/pub28/28apc_002.htm}{C1 Street Abbreviations}.
}
\usage{
usps_street
}
\description{
A tibble containing common street suffixes or suffix
abbreviations and their full equivalent. Useful as the \code{add_abbs} argument
of \code{\link[=normal_address]{normal_address()}}.
}
\keyword{datasets}
