% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_city}
\alias{valid_city}
\title{Many of the cities in America}
\format{A vectory of city names}
\usage{
valid_city
}
\description{
The \code{city} column of the \code{geo} tibble
}
\keyword{datasets}
