% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_name}
\alias{valid_name}
\title{US State Names}
\format{
A vector of state names (length 62).
}
\usage{
valid_name
}
\description{
The \code{state} column of the \code{usps_state} tibble.
}
\details{
Contains 12 more names than \link[datasets:state]{datasets::state.name}.
}
\keyword{datasets}
