% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_state}
\alias{valid_state}
\title{All 2-letter State Abbreviations}
\format{A vectory of 2-digit abbreviations}
\usage{
valid_state
}
\description{
The \code{state} column of the \code{geo} tibble
}
\keyword{datasets}
