% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_zip}
\alias{valid_zip}
\title{Almost all of the valid USA ZIP Codes}
\format{A vectory of 5-digit ZIP codes}
\usage{
valid_zip
}
\description{
The \code{zip} column of the \code{geo} tibble
}
\keyword{datasets}
