% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vt_contribs}
\alias{vt_contribs}
\title{Example Vermont Contributions}
\format{A tibble with 10 rows of 10 variables:
\describe{
\item{id}{A unique contribution ID.}
\item{cand}{The reporting candidate.}
\item{date}{The contribution date.}
\item{amount}{The contribution amount.}
\item{first}{The contributor's first name}
\item{last}{The contribution's last name}
\item{address}{The contributor's address}
\item{city}{The contributor's city}
\item{state}{The contributor's state}
\item{zip}{The contributor's zip}
...
}}
\usage{
vt_contribs
}
\description{
This randomly generated data frame shows some of the many
formatting problems politicians, journalists, and political scientists will
likely face when processing state level campaign finance data. Missing or
invalid values, duplicate rows, negative amounts, questionable dates, and
incredibly inconsistent string formatting. The functions in this package
have been created to help import, explore, and clean this kind of data to
increase public accountability.
}
\keyword{datasets}
