% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zipcodes}
\alias{zipcodes}
\title{US City, state, and ZIP}
\format{
A tibble with 44,336 rows of 3 variables:
\describe{
\item{city}{Normalized city name.}
\item{state}{Two letter state abbreviation.}
\item{zip}{Five-digit ZIP Code.}
...
}
}
\source{
Daniel Coven's federalgovernmentzipcodes.us web site and the
CivicSpace US ZIP Code Database written by Schuyler Erle
\href{mailto:schuyler@geocoder.us}{schuyler@geocoder.us}, 5 August 2004. Original CSV files available from
\url{http://federalgovernmentzipcodes.us/free-zipcode-database-Primary.csv}
}
\usage{
zipcodes
}
\description{
This tibble is the third version of a popular zipcodes database.
The original CivicSpace US ZIP Code Database was created by Schuyler Erle
using ZIP code gazetteers from the US Census Bureau from 1999 and 2000,
augmented with additional ZIP code information from the Census Bureau’s
TIGER/Line 2003 data set. The second version was published as the
\code{zipcode::zipcode} dataframe object. This version has dropped the latitude
and longitude, reorganized columns, and normalize the city values with
\code{\link[=normal_city]{normal_city()}}.
}
\keyword{datasets}
