% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrev_state.R
\name{abbrev_state}
\alias{abbrev_state}
\title{Abbreviate US State Names}
\usage{
abbrev_state(state, rm_nums = FALSE)
}
\arguments{
\item{state}{A vector of full state names}

\item{rm_nums}{Whether to remove numbers from abbrev (e.g., 1st).}
}
\value{
A vector of 2 letter state abbreviations
}
\description{
Abbreviate US State Names
}
\examples{
abbrev_state(c("Vermont", "District of Columbia", "France", "Maine 2nd"), rm_nums = TRUE)
}
