% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrev-state.R
\name{abbrev_state}
\alias{abbrev_state}
\alias{abbreviate_state}
\title{Abbreviate US State Names}
\usage{
abbrev_state(full, na_bad = FALSE, rm_nums = FALSE)
}
\arguments{
\item{full}{A full US state name character vector (e.g., "Vermont").}

\item{na_bad}{logical; Whether to return \code{NA} for invalid \code{full} names
(default \code{FALSE}).}

\item{rm_nums}{logical; Whether to remove numbers from \code{full} (e.g., 1st)
before abbreviating. Particularly useful with US Congressional districts.
Only numbers 1 through 20 are supported at this time (default \code{FALSE}).}
}
\value{
A vector of 2 letter state abbreviations.
}
\description{
This function is used to first normalize a \code{full} state name and
then return any matching abbreviation. This package works for all state
names in \code{state.name} plus 12 additional territories (e.g., "Puerto Rico").
}
\examples{
abbrev_state(full = state.name)
abbrev_state(full = c("new_mexico", "france"), na_bad = TRUE)
}
\seealso{
Other geographic normalization functions: \code{\link{abbrev_full}},
  \code{\link{expand_abbrev}}, \code{\link{expand_state}},
  \code{\link{normal_address}}, \code{\link{normal_city}},
  \code{\link{normal_state}}, \code{\link{normal_zip}}
}
\concept{geographic normalization functions}
