% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrev_state.R
\name{abbrev_state}
\alias{abbrev_state}
\title{AbbreviateUS State Names}
\usage{
abbrev_state(state)
}
\arguments{
\item{state}{A vector of full state names}
}
\value{
A vector of 2 letter state abbreviations
}
\description{
AbbreviateUS State Names
}
\examples{
abrev_state(c("Vermont", "District of Columbia", "new_hampshire", "VT"))
}
