% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abrev_state.R
\name{abrev_state}
\alias{abrev_state}
\title{AbbreviateUS State Names}
\usage{
abrev_state(state)
}
\arguments{
\item{state}{A vector of full state names}
}
\value{
A vector of 2 letter state abbreviations
}
\description{
AbbreviateUS State Names
}
