% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_files_new.R
\name{all_files_new}
\alias{all_files_new}
\title{Check if All Files Are New}
\usage{
all_files_new(path, glob = NULL, ...)
}
\arguments{
\item{path}{The path to a directory to check.}

\item{glob}{A pattern to search for files (e.g., "csv").}

\item{...}{Arguments to be passed to \code{fs::dir_ls()}}
}
\value{
A logical value indicating whether or not all files in the directory have a modification date equal to today.
}
\description{
Check if All Files Are New
}
