% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-date-usa.R
\name{col_date_usa}
\alias{col_date_usa}
\title{Parse USA date columns in readr functions}
\usage{
col_date_usa()
}
\value{
A \code{POSIXct} vector.
}
\description{
Parse dates with format MM/DD/YYYY. This function simply wraps around
\code{\link[readr:parse_datetime]{readr::col_date()}} with the \code{format} argument set to \code{"\%m/\%d/\%Y"}. Most US
campaign finance datasets use this format.
}
\examples{
readr::read_csv(file = "x\n11/09/2016", col_types = readr::cols(x = col_date_usa()))
}
