% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{count_diff}
\alias{count_diff}
\title{Count set difference}
\usage{
count_diff(x, y)
}
\arguments{
\item{x}{A vector to check.}

\item{y}{A vector to compare against.}
}
\value{
The number of \emph{unique} values of \code{x} not in \code{y}.
}
\description{
Find the length of the set of difference between \code{x} and \code{y} vectors.
}
\details{
\code{sum(x \%out\% y)}
}
\examples{
# Only unique values are checked
count_diff(c("VT", "NH", "ZZ", "ZZ", "ME"), state.abb)
}
\seealso{
Other Simple counting wrappers: 
\code{\link{count_in}()},
\code{\link{count_na}()},
\code{\link{count_out}()},
\code{\link{prop_in}()},
\code{\link{prop_na}()},
\code{\link{prop_out}()}
}
\concept{Simple counting wrappers}
