% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{count_na}
\alias{count_na}
\title{Count \code{NA}}
\usage{
count_na(x)
}
\arguments{
\item{x}{A vector to check.}
}
\value{
The sum of \code{x} that are \code{NA}
}
\description{
Count the values of a vector that are missing.
}
\details{
\code{sum(is.na(x))}
}
\examples{
count_na(c("VT", "NH", NA, "ME"))
}
\seealso{
Other Simple Counting Wrappers: \code{\link{count_in}},
  \code{\link{count_out}}, \code{\link{progress_table}},
  \code{\link{prop_in}}, \code{\link{prop_na}},
  \code{\link{prop_out}}
}
\concept{Simple Counting Wrappers}
