% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{count_out}
\alias{count_out}
\title{Count Out}
\usage{
count_out(x, y, na.rm = TRUE)
}
\arguments{
\item{x}{A vector to check.}

\item{y}{A vector to compare against.}

\item{na.rm}{logical; Should \code{NA} be ignored?}
}
\value{
The sum of \code{x} absent in \code{y}.
}
\description{
Find proportion of values of one vector not in another vector.
}
\details{
\code{sum(x \%out\% y)}
}
\examples{
count_out(c("VT", "NH", "ZZ", "ME"), state.abb)
}
\seealso{
Other Simple Counting Wrappers: \code{\link{count_diff}},
  \code{\link{count_in}}, \code{\link{count_na}},
  \code{\link{progress_table}}, \code{\link{prop_in}},
  \code{\link{prop_na}}, \code{\link{prop_out}}
}
\concept{Simple Counting Wrappers}
