% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_bar.R
\name{explore_plot}
\alias{explore_plot}
\title{Check for All New Files}
\usage{
explore_plot(data, var, flip = FALSE, head = 8, palette = "Set1",
  ...)
}
\arguments{
\item{data}{A data frame to glimpse.}

\item{fun}{A function to map to each column.}
}
\value{
A tibble with a row for every column and new columns with count and proportion.
}
\description{
Check for All New Files
}
\examples{
glimpse_fun(dplyr::storms, dplyr::n_distinct)
}
