% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_plot.R
\name{explore_plot}
\alias{explore_plot}
\title{Create Basic Barplots}
\usage{
explore_plot(data, var, flip = FALSE, nbar = 8, palette = "Dark2",
  ...)
}
\arguments{
\item{data}{The data frame to explore.}

\item{var}{A variable to plot.}

\item{flip}{logical; Whether plot should be flipped horizontally.}

\item{nbar}{The number of bars to plot. Always shows most common values.}

\item{palette}{The color palette passed to [ggplot2::scale_fill_brewer().}

\item{...}{Optional arguments passed to \code{\link[ggplot2:labs]{ggplot2::labs()}}.}
}
\value{
A bar plot.
}
\description{
This function simply wraps around \code{\link[ggplot2:geom_col]{ggplot2::geom_col()}} to take
a dataframe and categorical variable to return a simple barplot count.
}
\examples{
glimpse_fun(dplyr::storms, dplyr::n_distinct)
}
