% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_na.R
\name{flag_na}
\alias{flag_na}
\title{Flag Missing Values With New Column}
\usage{
flag_na(data, ...)
}
\arguments{
\item{data}{A data frame to flag.}

\item{...}{Arguments passed to \code{\link[dplyr:select]{dplyr::select()}} (needs to be at least \code{\link[dplyr:everything]{dplyr::everything()}}).}
}
\value{
A data frame with a new \code{na_flag} logical variable.
}
\description{
This function uses \code{\link[dplyr:mutate]{dplyr::mutate()}} to create a new \code{dupe_flag} logical variable
with \code{TRUE} values for any record duplicated more than once.
}
\examples{
flag_na(dplyr::starwars, gender)
}
