% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo}
\alias{geo}
\title{US City, State, and ZIPs}
\format{A tibble with 44,360 rows of 3 variables:
\describe{
\item{city}{City name}
\item{state}{state abbreviation}
\item{zip}{US ZIP code}
...
}}
\source{
Daniel Coven's federalgovernmentzipcodes.us web site and the CivicSpace US ZIP Code Database written by Schuyler Erle \href{mailto:schuyler@geocoder.us}{schuyler@geocoder.us}, 5 August 2004. Original CSV files available from \url{http://federalgovernmentzipcodes.us/free-zipcode-database-Primary.csv}, \url{http://www.boutell.com/zipcodes/}, and \url{http://mappinghacks.com/data/}.
}
\usage{
geo
}
\description{
A dataset containing all US ZIP codes alongside their city and state
}
\details{
The original author claims this dataset contains over 98% of all US ZIP codes and covers 99.9% of the population. It may contains up to 0.5% false positives.
}
\keyword{datasets}
