% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse-fun.R
\name{glimpse_fun}
\alias{glimpse_fun}
\title{Glimpse Column Count Functions}
\usage{
glimpse_fun(data, fun, print = TRUE)
}
\arguments{
\item{data}{A data frame to glimpse.}

\item{fun}{A function to map to each column.}

\item{print}{logical; Should all columns be printed as rows?}
}
\value{
A tibble with a row for every column and new columns with count and
proportion.
}
\description{
Apply a counting summary function like \code{\link[dplyr:n_distinct]{dplyr::n_distinct()}} or
\code{\link[=count_na]{count_na()}} to every column of a dataframe and return the results along
with a \emph{percentage} of that value.
}
\details{
Deprecated in favor or simpler \code{\link[=col_stats]{col_stats()}}.
}
