% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{\%out\%}
\alias{\%out\%}
\title{Inverted Match}
\usage{
x \%out\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched. Long vectors are supported.}

\item{table}{vector or NULL: the values to be matched against.}
}
\value{
Logical if X is not present in Y
}
\description{
\code{\%out\%} is an inverted version of the infix opperator \code{\%in\%}.
}
\details{
\code{\%out\%} is currently defined as \code{"\%out\%" <- function(x, table) match(x, table, nomatch = 0) == 0}
}
\examples{
c("A", "B", "3") \%out\% LETTERS
}
