% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_binary.R
\name{is_binary}
\alias{is_binary}
\title{Is Binary}
\usage{
is_binary(x)
}
\arguments{
\item{x}{A vector}
}
\value{
TRUE if only 2 unique values
}
\description{
Is Binary
}
\examples{
is_binary(x = c("Yes", "No"))
}
