% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-even.R
\name{is_even}
\alias{is_even}
\title{Check if even}
\usage{
is_even(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
logical; Whether the integer is even or odd.
}
\description{
Check if even
}
\examples{
is_even(1:10)
is_even(10L)
}
