% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-mdb.R
\name{mdb_tables}
\alias{mdb_tables}
\title{Get tables from Microsoft Access file}
\usage{
mdb_tables(file)
}
\arguments{
\item{file}{A path to a .mdb file.}
}
\value{
A vector of table names.
}
\description{
Wraps around the \code{mdb-tables} command line tool.
}
