% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_common.R
\name{most_common}
\alias{most_common}
\title{Find Most Common Values}
\usage{
most_common(x, n = 6)
}
\arguments{
\item{x}{A vector.}

\item{n}{Number of values to return.}
}
\value{
Sorted vector of \code{n} most common values.
}
\description{
From a character vector, which values are most common?
}
\examples{
most_common(iris$Species, n = 1)
}
