% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{na_in}
\alias{na_in}
\title{Remove In}
\usage{
na_in(x, y)
}
\arguments{
\item{x}{A vector to check.}

\item{y}{A vector to compare against.}
}
\value{
The vector \code{x} missing any values in \code{y}.
}
\description{
Remove the values of one vector that are in another vector.
}
\concept{Simple Counting Wrappers
na_in(c("VT", "NH", "ZZ", "ME"), state.abb)
na_in(1:10, seq(1, 10, 2))}
