% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_out.R
\name{na_out}
\alias{na_out}
\title{NA if Out}
\usage{
na_out(x, y)
}
\arguments{
\item{x}{A vector to check}

\item{y}{A vector to compare}
}
\value{
A vector \code{x} with NA for those not \code{\%in\%} \code{y}
}
\description{
NA if Out
}
\examples{
na_out(x = c("VT", "QC", "CA"), y = state.abb)
}
