% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{na_out}
\alias{na_out}
\title{Remove out}
\usage{
na_out(x, y)
}
\arguments{
\item{x}{A vector to check.}

\item{y}{A vector to compare against.}
}
\value{
The vector \code{x} missing any values not in \code{y}.
}
\description{
Set \code{NA} for the values of \code{x} that are \verb{\%out\%} of the vector \code{y}.
}
\concept{Simple counting wrappers
na_out(c("VT", "NH", "ZZ", "ME"), state.abb)
na_out(1:10, seq(1, 10, 2))}
