% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_address.R
\name{normal_address}
\alias{normal_address}
\title{Normalize US Street Addresses}
\usage{
normal_address(address, add_abbs = NULL, na = c(""), na_rep = FALSE)
}
\arguments{
\item{address}{A vector of street addresses}

\item{add_abbs}{An optional data frame of abbreviations and replacements}

\item{na}{A vector of values to make `NA`}

\item{na_rep}{If TRUE, make all single digit repeating strings `NA`}
}
\value{
A vector of normalized street addresses
}
\description{
Normalize US Street Addresses
}
