% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_address.R
\name{normal_address}
\alias{normal_address}
\title{Normalize Street Addresses}
\usage{
normal_address(address, add_abbs = NULL, na = c("", "NA"),
  na_rep = FALSE)
}
\arguments{
\item{address}{A vector of street addresses (ideally without city, state, and
ZIP code; you can use the \link{rx_state} and \link{rx_zip} regular expression
patterns to remove the later two).}

\item{add_abbs}{A two-column data frame like \link{usps_street}, with a full
suffix ("STREET") in the first column and abbreviations ("ST") in the
second. Replace all abbreviations with their full version.}

\item{na}{A vector of values to make \code{NA} (like \link{invalid_city}).}

\item{na_rep}{logical; If \code{TRUE}, replace all single digit (repeating)
strings with \code{NA}.}
}
\value{
A vector of normalized street addresses.
}
\description{
Return consistent version of a US Street Address using
\code{stringr::str_*()} functions. Letters are capitalized, hyphens and
underscores are replaced with whitespaces, other punctuation is removed,
and excess whitespace is trimed and squished. Optionally, street suffix
abbreviations ("AVE") can be replaced with their long form ("AVENUE").
Invalid addresses from a vector can be removed (possibly using
\link{invalid_city}) as well as single (repeating) character strings ("XXXXXX").
}
\examples{
normal_address("1600 Pennsylvania Ave NW", add_abbs = usps_street)
normal_address("12 e st main ave ste 209", add_abbs = usps_street)
}
