% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_city.R
\name{normal_city}
\alias{normal_city}
\title{Normalize City Names}
\usage{
normal_city(city, geo_abbs = NULL, st_abbs = NULL, na = c(""),
  na_rep = FALSE)
}
\arguments{
\item{city}{A vector of US city names.}

\item{geo_abbs}{A two-column data frame with geographic abbreviations ("MT") in the first column and replacement strings ("MOUNT") in the second.}

\item{st_abbs}{A vector of state abbreviations ("VT") to remove from the \emph{end} of city names ("STOWE VT").}

\item{na}{A vector of values to make \code{NA} (useful with the \code{campfin::na_city} vector).}

\item{na_rep}{If \code{TRUE}, make all single digit repeating strings \code{NA}.}
}
\value{
A vector of normalized city names.
}
\description{
Normalize City Names
}
\examples{
normal_city(
 city = c("Stowe, VT", "N/A", "Burlington", "ST JOHNSBURY", "XXXXXXXXX"),
 geo_abbs = tibble::tibble(abb = "ST", rep = "SAINT"),
 st_abbs = c("VT"),
 na = c("", "NA", "UNKNOWN"),
 na_rep = TRUE
)
}
