% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_zip.R
\name{normal_zip}
\alias{normal_zip}
\title{Normalize US ZIP Codes}
\usage{
normal_zip(zip, na = c("", "NA"), na_rep = FALSE)
}
\arguments{
\item{zip}{A vector of US ZIP codes.}

\item{na}{A vector of values to make \code{NA}.}

\item{na_rep}{If \code{TRUE}, make all single digit repeating strings (e.g., "00000") \code{NA}. This removes the valid but uncommon "22222", "44444", and "55555".}
}
\value{
A vector of normalized 5-digit ZIP codes.
}
\description{
Normalize US ZIP Codes
}
\examples{
normal_zip(
  zip = c("05672-5563", "N/A", "05401", "5819", "00000"),
  na = c("", "NA"),
  na_rep = TRUE
)
}
