% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{\%out\%}
\alias{\%out\%}
<<<<<<< HEAD
\title{Proportion In}
\usage{
... \%out\% NA
}
\arguments{
\item{x}{A vector to check.}

\item{y}{A vector to compare against.}
}
\value{
Logical if X is not present in Y
}
\description{
Proportion In
}
\examples{
c("A", "B", "0") \%out\% LETTERS
}
=======
\title{Out opperator}
\usage{
x %out% y
}
\description{
Out opperator
}
\keyword{internal}
>>>>>>> b9f842d7addc2db1b363cd20b39d4705201d0bf0
