% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-abbrev.R
\name{path.abbrev}
\alias{path.abbrev}
\title{Abbreviate a file path}
\usage{
path.abbrev(path, dir = fs::path_wd())
}
\arguments{
\item{path}{Character vector containing one or more full paths.}

\item{dir}{The directory to replace with \code{~}. Defaults to \code{\link[fs:path]{fs::path_wd()}}.}
}
\value{
Abbreviated file paths.
}
\description{
This is an inverse of \code{\link[=path.expand]{path.expand()}}, which replaces the home directory or
project directory with a tilde.
}
\examples{
print(fs::path_wd("test"))
path.abbrev(fs::path_wd("test"))
}
