% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress-table.R
\name{progress_table}
\alias{progress_table}
\title{Create a progress table}
\usage{
progress_table(..., compare)
}
\arguments{
\item{...}{Any number of vectors to check.}

\item{compare}{A vector to compare each of \code{...} against. Useful
with \link{valid_zip}, \link{valid_state} (\link{valid_name}), or
\link{valid_city}.}
}
\value{
A table with a row for each vector in \code{...}.
}
\description{
Create a tibble with rows for each stage of normalization and columns for the
various statistics most useful in assessing the progress of each stage.
}
\examples{
progress_table(state.name, toupper(state.name), compare = valid_name)
}
\concept{Simple Counting Wrappers}
