% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_io.R
\name{prop_in}
\alias{prop_in}
\title{Proportion In}
\usage{
prop_in(x, y, na.rm = TRUE)
}
\arguments{
\item{x}{A vector to check.}

\item{y}{A vector to compare against.}

\item{na.rm}{Should NA values be ignored.}
}
\value{
The proprtion of X present in Y.
}
\description{
Proportion In
}
\examples{
prop_in(c("VT", "NH", "ZZ", "ME"), state.abb)
}
