% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_fun.R
\name{prop_na}
\alias{prop_na}
\title{Count NA Proportion}
\usage{
prop_na(x, format = FALSE)
}
\arguments{
\item{x}{A vector to count \code{NA}}

\item{format}{Should the output use \code{scales::percent()}?}
}
\value{
The proportion of \code{NA} values in a vector
}
\description{
Wrap around \code{mean(is.na())}
}
\examples{
prop_na(dplyr::starwars$gender)
}
