% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{prop_na}
\alias{prop_na}
\title{Proportion \code{NA}}
\usage{
prop_na(x)
}
\arguments{
\item{x}{A vector to check.}
}
\value{
The sum of \code{x} absent in \code{y}.
}
\description{
Find the proportion of values of a vector that are missing.
}
\details{
\code{mean(is.na(x))}
}
\examples{
prop_na(c("VT", "NH", NA, "ME"))
}
\seealso{
Other Simple Counting Wrappers: \code{\link{count_in}},
  \code{\link{count_na}}, \code{\link{count_out}},
  \code{\link{prop_in}}, \code{\link{prop_out}}
}
\concept{Simple Counting Wrappers}
