% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_io.R
\name{prop_out}
\alias{prop_out}
\title{Proportion Out}
\usage{
prop_out(x, y, na.rm = TRUE)
}
\arguments{
\item{x}{A vector to check.}

\item{y}{A vector to compare against.}

\item{na.rm}{Should \code{NA} be ignored?}
}
\value{
The proprtion of X \emph{not} present in Y.
}
\description{
Proportion Out
}
\examples{
prop_in(c("VT", "NH", "ZZ", "ME"), state.abb)
}
