% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rx_state}
\alias{rx_state}
\title{State regex}
\format{
A character string (length 1).
}
\usage{
rx_state
}
\description{
The regex string to extract state string preceding ZIP code.
}
\keyword{datasets}
