% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rx_zip}
\alias{rx_zip}
\title{ZIP code regex}
\format{
A character string (length 1).
}
\usage{
rx_zip
}
\description{
The regex string to extract ZIP code from the end of address.
}
\keyword{datasets}
