% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-truncate.R
\name{scale_x_truncate}
\alias{scale_x_truncate}
\title{Truncate x-axis labels}
\usage{
scale_x_truncate(n = 20, ...)
}
\arguments{
\item{n}{The maximum width of string passed to \code{\link[stringr:str_trunc]{stringr::str_trunc()}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_discrete]{ggplot2::scale_x_discrete()}}.}
}
\description{
Truncate the labels of a plot's discrete x-axis labels so that the text does
not overflow and collide with other bars.
}
