% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-dist.R
\name{str_dist}
\alias{str_dist}
\title{Calculate string distance}
\usage{
str_dist(a, b, method = "osa", ...)
}
\arguments{
\item{a}{\code{R} object (target); will be converted by \code{\link[base:character]{base::as.character()}}.}

\item{b}{\code{R} object (source); will be converted by \code{\link[base:character]{base::as.character()}}.}

\item{method}{Method for distance calculation. The default is "osa."}

\item{...}{Other arguments passed to \code{\link[stringdist:stringdist]{stringdist::stringdist()}}.}
}
\value{
The distance between string \code{a} and string \code{b}.
}
\description{
This function wraps around \code{\link[stringdist:stringdist]{stringdist::stringdist()}}.
}
\examples{
str_dist(a = "BRULINGTN", b = "BURLINGTON")
}
