% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_files_new.R
\name{this_file_new}
\alias{this_file_new}
\title{Check if This File is New}
\usage{
this_file_new(path)
}
\arguments{
\item{path}{The path to a file to check.}
}
\value{
logical; whether the files has a modification date equal to today.
}
\description{
This function tests whether a single file has a modification
date equal to the system date. Useful when repeatedly running code with a
lengthy download stage. Many state databases are updated daily, so new data
can be helpful but not always neccesary. Set this function in an \code{if}
statement.
}
\examples{
temp <- tempfile()
if (!this_file_new(temp)) {
  download.file(
    url = "http://212.183.159.230/5MB.zip",
    destfile = temp
  )
}
}
