% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files-new.R
\name{this_file_new}
\alias{this_file_new}
\title{Check if a single file is new}
\usage{
this_file_new(path)
}
\arguments{
\item{path}{The path to a file to check.}
}
\value{
logical; Whether the file has a modification date equal to today.
}
\description{
This function tests whether a single file has a modification date equal to
the system date. Useful when repeatedly running code with a lengthy download
stage. Many state databases are updated daily, so new data can be helpful but
not always necessary. Set this function in an \code{if} statement.
}
\examples{
tmp <- tempfile()
this_file_new(tmp)
}
