% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usps_city}
\alias{usps_city}
\title{USPS City Abbreviations}
\format{
A tibble with 154 rows of 2 variables:
\describe{
\item{full}{Primary Street Suffix}
\item{abb}{Commonly Used Street Suffix or Abbreviation}
...
}
}
\source{
USPS Appendix C1,
\href{https://pe.usps.com/text/pub28/28apc_002.htm}{Street Abbreviations}
}
\usage{
usps_city
}
\description{
A curated and edited subset of \link{usps_street} containing the
USPS abbreviations found in city names. Useful as the \code{geo_abbs} argument
of \code{\link[=normal_city]{normal_city()}}.
}
\keyword{datasets}
