% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_city}
\alias{valid_city}
\title{US City Names}
\format{
A sorted vector of unique city names (length 19,083).
}
\usage{
valid_city
}
\description{
The \code{city} column of the \code{zipcodes} tibble.
}
\keyword{datasets}
