% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_state}
\alias{valid_state}
\title{US State Abbreviations}
\format{
A vector of 2-digit abbreviations (length 62).
}
\usage{
valid_state
}
\description{
The \code{abb} column of the \code{usps_state} tibble.
}
\keyword{datasets}
